+++ Adapter/PlentymarketsAdapter/ResponseParser/Product/ProductResponseParser.php
@@ -21,8 +21,11 @@ use PlentyConnector\Connector\TransferObject\TransferObjectInterface;
 use PlentyConnector\Connector\TransferObject\VatRate\VatRate;
 use PlentyConnector\Connector\ValueObject\Attribute\Attribute;
 use PlentyConnector\Connector\ValueObject\Translation\Translation;
+use PlentymarketsAdapter\Client\ClientInterface;
 use PlentymarketsAdapter\Helper\VariationHelperInterface;
 use PlentymarketsAdapter\PlentymarketsAdapter;
+use PlentymarketsAdapter\ReadApi\Item\Property\Group as PropertyGroupApi;
+use PlentymarketsAdapter\ReadApi\Item\Property\Name as PropertyNameApi;
 use PlentymarketsAdapter\ResponseParser\Product\Image\ImageResponseParserInterface;
 use PlentymarketsAdapter\ResponseParser\Product\Variation\VariationResponseParserInterface;
 use Psr\Log\LoggerInterface;
@@ -59,13 +62,24 @@ class ProductResponseParser implements ProductResponseParserInterface
      */
     private $variationHelper;
 
+    /**
+     * @var PropertyGroupApi
+     */
+    private $itemsPropertiesGroupNamesApi;
+
+    /**
+     * @var PropertyNameApi
+     */
+    private $itemsPropertiesNamesApi;
+
     public function __construct(
         ConfigServiceInterface $configService,
         IdentityServiceInterface $identityService,
         LoggerInterface $logger,
         ImageResponseParserInterface $imageResponseParser,
         VariationResponseParserInterface $variationResponseParser,
-        VariationHelperInterface $variationHelper
+        VariationHelperInterface $variationHelper,
+        ClientInterface $client
     ) {
         $this->configService = $configService;
         $this->identityService = $identityService;
@@ -73,6 +87,9 @@ class ProductResponseParser implements ProductResponseParserInterface
         $this->imageResponseParser = $imageResponseParser;
         $this->variationResponseParser = $variationResponseParser;
         $this->variationHelper = $variationHelper;
+
+        $this->itemsPropertiesGroupsNamesApi = new PropertyGroupApi($client);
+        $this->itemsPropertiesNamesApi = new PropertyNameApi($client);
     }
 
     /**
@@ -461,10 +478,63 @@ class ProductResponseParser implements ProductResponseParserInterface
                 continue;
             }
 
+            $backendName = $property['property']['backendName'];
+
             $values = [];
             $translations = [];
 
-            if ($property['property']['valueType'] === 'text') {
+            if ($property['property']['valueType'] === 'empty') {
+                $propertyGroupNames = $this->itemsPropertiesGroupsNamesApi->findOne($property['property']['propertyGroupId']);
+
+                if (empty($propertyGroupNames[0]['name'])) {
+                    continue;
+                }
+
+                $backendName = $propertyGroupNames[0]['name'];
+
+                foreach ($propertyGroupNames as $name) {
+                    $languageIdentifier = $this->identityService->findOneBy([
+                        'adapterIdentifier' => $name['lang'],
+                        'adapterName' => PlentymarketsAdapter::NAME,
+                        'objectType' => Language::TYPE,
+                    ]);
+
+                    if (null === $languageIdentifier) {
+                        continue;
+                    }
+
+                    $translations[] = Translation::fromArray([
+                        'languageIdentifier' => $languageIdentifier->getObjectIdentifier(),
+                        'property' => 'name',
+                        'value' => $name['name'],
+                    ]);
+                }
+
+                $propertyNames = $this->itemsPropertiesNamesApi->findOne($property['property']['id']);
+
+                foreach ($propertyNames as $name) {
+                    $languageIdentifier = $this->identityService->findOneBy([
+                        'adapterIdentifier' => $name['lang'],
+                        'adapterName' => PlentymarketsAdapter::NAME,
+                        'objectType' => Language::TYPE,
+                    ]);
+
+                    if (null === $languageIdentifier) {
+                        continue;
+                    }
+
+                    $valueTranslations[] = Translation::fromArray([
+                        'languageIdentifier' => $languageIdentifier->getObjectIdentifier(),
+                        'property' => 'value',
+                        'value' => $name['name'],
+                    ]);
+                }
+
+                $values[] = Value::fromArray([
+                    'value' => (string) $property['property']['backendName'],
+                    'translations' => $valueTranslations,
+                ]);
+            } elseif ($property['property']['valueType'] === 'text') {
                 if (empty($property['names'][0]['value'])) {
                     continue;
                 }
@@ -556,7 +626,7 @@ class ProductResponseParser implements ProductResponseParserInterface
             }
 
             $result[] = Property::fromArray([
-                'name' => $property['property']['backendName'],
+                'name' => $backendName,
                 'values' => $values,
                 'translations' => $translations,
             ]);