+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - fixed duplicate product seo categories (@lacodimizer)
 - the media category sync was called multiple times instead of only one time (@jochenmanz)
 - skip order without customer
+- set correct bundle position
+- fix product meta title and meta description translation
 
 ### Changed
 - translated variation configurator 
@@ -327,6 +329,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - transfer age restriction as attribute to shopware
+- add item properties with type "empty"
 
 ## [4.6.0]
 ### Fixed